# backend/src/main/java/com/yoursay/backend/domain/OpinionRequest.java
package com.yoursay.backend.domain;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class OpinionRequest {
    private String email;
    private Integer bill_id;
    private String opinion;
}