# backend/swagger/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "YourSay API",
    "description": "API documentation for YourSay backend services.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/users/preferences": {
      "put": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "updatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "verifyAndAddUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/send-verification": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "sendVerificationText",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "checkLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/vote": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "addVoteOnBill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/opinion": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "addOpinionOnBill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpinionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/local": {
      "post": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "addLocalLegislation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalLegislationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/state/{state}": {
      "get": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "fetchStateLegislation",
        "parameters": [
          {
            "name": "state",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Legislation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/random/{zipcode}/{state}": {
      "get": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "fetchRandomLegislation",
        "parameters": [
          {
            "name": "zipcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Legislation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/local/{zipcode}": {
      "get": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "fetchLocalLegislation",
        "parameters": [
          {
            "name": "zipcode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Legislation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/federal": {
      "get": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "fetchFederalLegislation",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Legislation"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/legislation/daily-fetch": {
      "get": {
        "tags": [
          "legislation-controller"
        ],
        "operationId": "fetchDailyLegislation",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PreferencesRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "zipcode": {
            "type": "string"
          },
          "preferences": {
            "type": "string"
          }
        }
      },
      "UserRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "zipcode": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "preferences": {
            "type": "string"
          },
          "verificationCode": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "VerificationRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "zipcode": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "preferences": {
            "type": "string"
          },
          "accessGranted": {
            "type": "boolean"
          }
        }
      },
      "VoteRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "bill_id": {
            "type": "integer",
            "format": "int32"
          },
          "vote": {
            "type": "boolean"
          }
        }
      },
      "OpinionRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "bill_id": {
            "type": "integer",
            "format": "int32"
          },
          "opinion": {
            "type": "string"
          }
        }
      },
      "LocalLegislationRequest": {
        "type": "object",
        "properties": {
          "zipcode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        }
      },
      "Legislation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "bill_id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "billLevel": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipcode": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "billDate": {
            "type": "string"
          },
          "category": {
            "type": "string"
          }
        }
      }
    }
  }
}